requirements.txt
# ScriptForge - AI Video Script Generator
# Core dependencies

# Web framework
streamlit>=1.28.0

# AI API clients
requests>=2.31.0

# Document export
python-docx>=1.1.0

# Environment configuration
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0  # For potential future features

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0

# Testing
pytest>=7.4.0